Refactor marketing page to integrate new landing page components and enhance animations. Remove unused variables and improve layout styling for better user experience.
