+++ src/routes/(marketing)/marketing-layout.css
@@ -160,3 +160,24 @@ html:has(.mkt) {
 	padding-left: max(var(--mkt-nav-inline), env(safe-area-inset-left, 0px));
 	padding-right: max(var(--mkt-nav-inline), env(safe-area-inset-right, 0px));
 }
+
+.mkt .landing-reveal {
+	opacity: 0;
+	transform: translateY(22px);
+	transition:
+		opacity 0.75s cubic-bezier(0.16, 1, 0.3, 1),
+		transform 0.75s cubic-bezier(0.16, 1, 0.3, 1);
+}
+
+.mkt .landing-reveal.is-revealed {
+	opacity: 1;
+	transform: translateY(0);
+}
+
+@media (prefers-reduced-motion: reduce) {
+	.mkt .landing-reveal {
+		opacity: 1 !important;
+		transform: none !important;
+		transition: none !important;
+	}
+}
